fix: block sync stops too soon on first load
The first load of block sync is stopping at the point of failing to get a header. This is 2-fold - on first load, it needs to fail of a block header isn’t available and not store progress. It also needs to find an rpc client for the parent header, not the current one.

fix: block sync progress not working during first load