fix: keep ssh private key as file path
The ssh private key is now only stored on the file system and the file path is passed around.

The closing of ssh connections is also simplified to avoid repeated invalid closing of connections due to closing on drop asynchonously, which could grab the wrong connection (eg, the newly created one).