feat: run local docker
Local Docker: the local docker runs a vm locally that allows us to run the same installer/linux based commands, but stores files on the host machine.

Specific Port: a custom server (and docker) can specify a custom ssh port if desired.

Status Server: the installer shell scripts were moved to a simple nodejs script since we no longer know what the compose installation will be and the scripts weren’t always accessible to run because of the way docker mounts the docker inside the other docker host.

Autostart/Always Awake: This feature adds an ability to run on a local docker machine, which will install the commander app as a keepawake app and check that docker is running on boot of the app. It also then makes Commander autostart.

Port Check: This configuration required changes to how ports are allocated (they’re not going to always be open on a local machine).

- Retry SSH: there was some flakiness in ssh availability if it was tried too quickly which was not recoverable.
- Installer: now correctly waits for the bot to be ready before leaving installing cloud machine page